fix: small design fixes and improvements (#43)
* fix: set a max height to the page toc
When the page has a lot of heading, because of `position: sticky`, the
ToC was truncated on large viewports. By settings a max height and
`overflow-y: auto`, the ToC is now always readable.
* fix: always align line numbers
When mixing diff tokens with line numbers without diff, some times the
line numbers was misaligned. It's not ideal, by increasing the min
width fixes that.
* fix: do not use homepage title as section title
I was doing that because Markdown can't map headings to components but
since we now support MDX, I think it's more confusing than helpful to
use the title as section title.
* fix: make identity card items centered
* fix: remove updatedOn property for bookmarks
* feat: add an isQuote property for bookmarks
I usually copy/paste some text from the link as description, so it
should be set as a quote and not as a content I wrote.
* fix: make sure img in links inherit from link border radius
* fix: do not display tag body if empty
* fix: expect an object instead of a string for repository property
* fix: remove an unwanted console.log
* feat: add an isArchived property for projects
* fix: use the correct pagination route for blog posts
* fix: do not set blog posts and guides on multiple columns
* fix: improve pagination position on the page